more efficient combinations: :

+++ include/finmath/OptionPricing/binomial_tree.h
@@ -3,6 +3,6 @@
 
 #include "options_pricing_types.h"
 
-double binomial_option_pricing(OptionType type, double S0, double K, double T, double r, double sigma, int N);
+double binomial_option_pricing(OptionType type, double S0, double K, double T, double r, double sigma, long N);
 
 #endif
